requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bardapi==0.1.31
beautifulsoup4==4.12.2
browser-cookie3==0.19.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
deep-translator==1.11.4
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
googleapis-common-protos==1.60.0
grpcio==1.56.2
grpcio-status==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
jwt==1.3.1
langdetect==1.0.9
lz4==4.3.2
multidict==6.0.4
openai==0.27.8
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.66.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2